перенес создание меню из фрагмента в активити
